Squashed On-The-Go mode (1/2)
On-The-Go turns on the camera of your phone and streams it to an overlay,
which is drawn over everything.
this makes you feel like your display is transparent, as you see the content
behind it.

its really useful when you are using your phone while walking, thats why
we called it "On-The-Go"

allows to change transparency of the overlay via the notification.

Modified parts of the code to make working for FusionSP
Enable/disable with QS tile, longpress on tile for options.

Credits: NamelessRom

Signed-off-by: dragon <dragon7780@hotmail.nl>

Conflicts:
	core/java/android/provider/Settings.java
	core/res/res/values/fusion_symbols.xml
	packages/SystemUI/res/values/slim_strings.xml
	packages/SystemUI/src/com/android/systemui/statusbar/phone/QuickSettingsController.java
	policy/src/com/android/internal/policy/impl/GlobalActions.java